services/gpu/requirements.txt
# GPU Service Requirements
# For RTX 4090 with CUDA support

# Core
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers & Models
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0  # For 4-bit quantization
sentencepiece>=0.1.99  # For tokenizers

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Async & Queue
redis>=5.0.0
aiofiles>=23.2.1

# Utilities
numpy>=1.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Monitoring
psutil>=5.9.0

# For CLIP
ftfy>=6.1.0
regex>=2023.10.3

# For LLaVA
einops>=0.7.0
timm>=0.9.0